Update resource_web_list.php
Updated to allow segment2 to control the active tab that is set.  Simply set the segment1 of the URL to be a name that you want to track.  Then if you want to have have a tab set to be active (other than "all") simply add that as the segment2.
page-uri/segment1/segment2
page-uri/trackpage/activetab